+++ strategy/from_json/default.ts
@@ -1,7 +1,6 @@
 // Copyright 2018-2021 Gamebridge.ai authors. All rights reserved. MIT license.
-import { JSONValue } from "../../serializable.ts";
 
 /** typed function */
-export function fromJSONDefault(value: JSONValue): any {
+export function fromJSONDefault<T>(value: T): T {
   return value;
 }