+++ strategy/to_json/from_object_containing_test.ts
@@ -13,7 +13,6 @@ test({
       @SerializeProperty()
       someClassProp = "test";
     }
-
     class Test extends Serializable {
       @SerializeProperty({ toJSONStrategy: fromObjectContaining() })
       test: { [k: string]: SomeClass } = { testing: new SomeClass() };