+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to
 
 - removes array as an strategy, use `composeStrategy` instead
 - fmt with deno@1.10.3
-
+- turn off unstable code coverage
 
 ## [v1.3.1] - 2021-03-14